Rename from MaxEnt to MDP
